pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "compass"
version = "0.1.0"
description = "Cross-solver Optical Modeling Platform for Advanced Sensor Simulation"
readme = {text = "COMPASS: Cross-solver Optical Modeling Platform for Advanced Sensor Simulation", content-type = "text/plain"}
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "COMPASS Team" },
]

dependencies = [
    "numpy>=1.24",
    "torch>=2.0",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "pydantic>=2.0",
    "h5py>=3.8",
    "matplotlib>=3.7",
    "scipy>=1.10",
    "pyyaml>=6.0",
    "tqdm>=4.60",
]

[project.optional-dependencies]
rcwa = [
    "torcwa>=0.1.4",
]
fdtd = [
    "fdtd>=0.3.0",
]
viz = [
    "pyvista>=0.39",
    "plotly>=5.14",
]
all = [
    "compass[rcwa,fdtd,viz]",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["compass"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
